changing buffer size parameter to size_t type

+++ engine/platform/opengl/opengl_buffer.cpp
@@ -6,14 +6,14 @@
 || VERTEX BUFFER                                                   ||
 \*=================================================================*/
 
-OpenGLVertexBuffer::OpenGLVertexBuffer(uint32_t size)
+OpenGLVertexBuffer::OpenGLVertexBuffer(size_t size)
 {
 	glCreateBuffers(1, &m_renderer_id);
 	glBindBuffer(GL_ARRAY_BUFFER, m_renderer_id);
 	glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_DYNAMIC_DRAW);
 }
 
-OpenGLVertexBuffer::OpenGLVertexBuffer(float* vertices, uint32_t size)
+OpenGLVertexBuffer::OpenGLVertexBuffer(float* vertices, size_t size)
 {
 	glCreateBuffers(1, &m_renderer_id);
 	glBindBuffer(GL_ARRAY_BUFFER, m_renderer_id);
@@ -35,7 +35,7 @@ void OpenGLVertexBuffer::unbind() const
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
 }
 
-void OpenGLVertexBuffer::push_data(const void* data, uint32_t size)
+void OpenGLVertexBuffer::push_data(const void* data, size_t size)
 {
 	glBindBuffer(GL_ARRAY_BUFFER, m_renderer_id);
 	glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
